linux-software/requirements.txt
pyserial==2.6
pyusb==1.0.0a3
wsgiref==0.1.2
